chore: remove batchSearchRecord specificy APIs from `datashare-tasks`
