+++ datashare-app/src/main/java/org/icij/datashare/CliApp.java
@@ -1,12 +1,12 @@
 package org.icij.datashare;
 
-import org.icij.datashare.asynctasks.TaskManager;
 import org.icij.datashare.cli.CliExtensionService;
 import org.icij.datashare.cli.spi.CliExtension;
 import org.icij.datashare.mode.CommonMode;
 import org.icij.datashare.tasks.ArtifactTask;
 import org.icij.datashare.tasks.CreateNlpBatchesFromIndex;
 import org.icij.datashare.tasks.DatashareTaskFactory;
+import org.icij.datashare.tasks.DatashareTaskManager;
 import org.icij.datashare.tasks.DeduplicateTask;
 import org.icij.datashare.tasks.EnqueueFromIndexTask;
 import org.icij.datashare.tasks.ExtractNlpTask;
@@ -64,7 +64,7 @@ class CliApp {
     }
 
     private static void runTaskWorker(CommonMode mode, Properties properties) throws Exception {
-        TaskManager taskManager = mode.get(TaskManager.class);
+        DatashareTaskManager taskManager = mode.get(DatashareTaskManager.class);
         DatashareTaskFactory taskFactory = mode.get(DatashareTaskFactory.class);
         Indexer indexer = mode.get(Indexer.class);
