+++ datashare-app/src/main/java/org/icij/datashare/tasks/CreateNlpBatchesFromIndex.java
@@ -52,7 +52,7 @@ public class CreateNlpBatchesFromIndex extends DefaultTask<List<String>> impleme
 
     private final User user;
     private volatile Thread taskThread;
-    private final TaskManager taskManager;
+    private final DatashareTaskManager taskManager;
     private final String searchQuery;
     private final Map<String, Object> batchTaskArgs;
     private final Pipeline.Type nlpPipeline;
@@ -72,7 +72,7 @@ public class CreateNlpBatchesFromIndex extends DefaultTask<List<String>> impleme
 
     @Inject
     public CreateNlpBatchesFromIndex(
-        final TaskManager taskManager, final Indexer indexer, @Assisted Task<LinkedList<String>> taskView,
+        final DatashareTaskManager taskManager, final Indexer indexer, @Assisted Task<LinkedList<String>> taskView,
         @Assisted final Function<Double, Void> ignored
     ) {
         this.user = taskView.getUser();
