+++ datashare-app/src/main/java/org/icij/datashare/web/StatusResource.java
@@ -14,8 +14,8 @@ import net.codestory.http.constants.HttpStatus;
 import net.codestory.http.payload.Payload;
 import org.icij.datashare.PropertiesProvider;
 import org.icij.datashare.Repository;
-import org.icij.datashare.asynctasks.TaskManager;
 import org.icij.datashare.openmetrics.StatusMapper;
+import org.icij.datashare.tasks.DatashareTaskManager;
 import org.icij.datashare.text.indexing.Indexer;
 
 import java.io.IOException;
@@ -26,10 +26,10 @@ public class StatusResource {
     private final PropertiesProvider propertiesProvider;
     private final Repository repository;
     private final Indexer indexer;
-    private final TaskManager taskManager;
+    private final DatashareTaskManager taskManager;
 
     @Inject
-    public StatusResource(PropertiesProvider propertiesProvider, Repository repository, Indexer indexer, TaskManager taskManager) {
+    public StatusResource(PropertiesProvider propertiesProvider, Repository repository, Indexer indexer, DatashareTaskManager taskManager) {
         this.propertiesProvider = propertiesProvider;
         this.repository = repository;
         this.indexer = indexer;
