+++ datashare-app/src/test/java/org/icij/datashare/tasks/CreateNlpBatchesFromIndexTest.java
@@ -34,7 +34,7 @@ public class CreateNlpBatchesFromIndexTest {
     public static ElasticsearchRule es = new ElasticsearchRule();
     private static final ElasticsearchIndexer indexer = new ElasticsearchIndexer(es.client, new PropertiesProvider())
         .withRefresh(Refresh.True);
-    private static TaskManager taskManager;
+    private static DatashareTaskManager taskManager;
 
     @Before
     public void setUp() {
