+++ datashare-app/src/test/java/org/icij/datashare/web/StatusResourceTest.java
@@ -2,7 +2,7 @@ package org.icij.datashare.web;
 
 import org.icij.datashare.PropertiesProvider;
 import org.icij.datashare.Repository;
-import org.icij.datashare.asynctasks.TaskManager;
+import org.icij.datashare.tasks.DatashareTaskManager;
 import org.icij.datashare.test.DatashareTimeRule;
 import org.icij.datashare.text.indexing.Indexer;
 import org.icij.datashare.web.testhelpers.AbstractProdWebServerTest;
@@ -22,7 +22,8 @@ public class StatusResourceTest extends AbstractProdWebServerTest {
     @Rule public DatashareTimeRule time = new DatashareTimeRule("2020-06-30T15:31:00Z");
     @Mock Repository repository;
     @Mock Indexer indexer;
-    @Mock TaskManager taskManager;
+    @Mock
+    DatashareTaskManager taskManager;
 
     @Before
     public void setUp() {
